# docs/openapi.json
{
...
  "info": {
    "description": "Submit workflows to be run on REANA Cloud",
    "title": "REANA Server",
    "version": "0.95.0a6"
  },
  "paths": {
    "/api/.well-known/openid-configuration": {
      "get": {
        "description": "Relays the OIDC discovery document of the deployment's trusted issuer, extended with the public client id that reana-client must use for the device authorization grant. This lets clients discover the identity provider knowing only the REANA URL.",
        "operationId": "get_openid_configuration",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request succeeded. The response contains the issuer's OpenID configuration and the REANA CLI client id.",
            "schema": {
              "properties": {
                "authorization_endpoint": {
                  "type": "string"
                },
                "device_authorization_endpoint": {
                  "type": "string"
                },
                "issuer": {
                  "type": "string"
                },
                "jwks_uri": {
                  "type": "string"
                },
                "reana_cli_client_id": {
                  "type": "string"
                },
                "reana_client_id": {
                  "type": "string"
                },
                "token_endpoint": {
                  "type": "string"
                },
                "userinfo_endpoint": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "The identity provider integration is not correctly configured.",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "502": {
            "description": "Request failed. The issuer's OpenID configuration could not be fetched.",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [],
        "summary": "Get the trusted issuer's OpenID configuration."
      }
    },
...
    "/api/login": {
      "get": {
        "description": "Redirects the browser to the trusted issuer's authorization endpoint (authorization code flow with PKCE). On completion the issuer redirects back to the OAuth callback, which establishes the cookie-based session. Returns 404 when the BFF login is disabled.",
        "operationId": "bff_login",
        "parameters": [
          {
            "description": "Relative URL to return to after login.",
            "in": "query",
            "name": "next",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the issuer's authorization endpoint."
          },
          "404": {
            "description": "BFF login is not enabled on this deployment."
          },
          "500": {
            "description": "The identity provider integration is not correctly configured."
          },
          "502": {
            "description": "The issuer's endpoints could not be resolved."
          }
        },
        "security": [],
        "summary": "Start the browser login flow."
      }
    },
    "/api/logout": {
      "post": {
        "description": "Deletes the server-side session (refresh token), clears the authentication cookies and returns the issuer's RP-initiated logout URL for the web application to navigate to.",
        "operationId": "bff_logout",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Session ended. The response contains the issuer logout URL.",
            "schema": {
              "properties": {
                "logout_url": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "No session cookie present."
          },
          "403": {
            "description": "CSRF validation failed."
          },
          "404": {
            "description": "Browser login is not enabled."
          },
          "503": {
            "description": "Browser session storage is temporarily unavailable."
          }
        },
        "security": [],
        "summary": "End the browser session."
      }
    },
    "/api/oauth/callback": {
      "get": {
        "description": "Handles the issuer's redirect: validates the OAuth state, exchanges the authorization code for tokens, provisions/links the REANA user, stores the refresh token server-side and sets the authentication cookies.",
        "operationId": "bff_oauth_callback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect back to the web application."
          },
          "403": {
            "description": "OAuth state validation failed."
          },
          "404": {
            "description": "Browser login is not enabled."
          },
          "500": {
            "description": "The identity provider integration is not correctly configured."
          },
          "502": {
            "description": "Token exchange with the issuer failed."
          },
          "503": {
            "description": "Browser session storage is temporarily unavailable."
          }
        },
        "security": [],
        "summary": "Complete the browser login flow."
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "OIDC access token using the `Bearer <token>` scheme.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}